Improved the memory consumption for the kWLinv() method.
Improved the memory consumption for the kWLinv() method (Still left to do for the kWL() method).
Changed the copyright notice to 2021.
Added a hashCode() function for Color and InitialColor classes, and changed the equals method's signature to take an Object arg so that it is used correctly by HashSets and HashMap.
Changed the color generation in kWLinv() so that memory consumption is greatly reduced, the downside is that all colors have to be generated twice. Running the method on a single group of order 512 now works for the first time on my PC.
Testing improving the performance in kWLinv() by using HashMap and HashSets instead of (Indexed) TreeMaps and TreeSets. The performance should be improved greatly, but the performance increase (from testing) doesn't seem to be as great as I would have hoped.
Added OutputGroupsStartingFromOrder512.bat and OutputGroupsStartingFromOrder512.g.